refactor: Separate specific from generic components
