+++ Spotify/src/screens/library/LibraryTemp.tsx
@@ -4,7 +4,7 @@ import {libHeaderVertPadding} from '../../constants/libraryHeader/libraryHeader'
 import {Padding} from '../../constants/styles/styleProps';
 import SnackbarUndoView from '../../components/snackbar/SnackbarUndoView';
 import ScreenView from '../../components/container/ScreenView';
-import FavoriteSongView from '../../components/favoriteSong/view/FavoriteSongView';
+import FavoriteSongView from './components/favoriteSong/view/FavoriteSongView';
 import LibraryHeader from '../../components/headers/LibraryHeader';
 
 type Props = {style?: StyleProp<ViewStyle> | undefined};
